moved bank-api.js into lib
